pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.1"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
line-bot-sdk = "^2.4.2"
openai = "^0.27.2"
langchain = "^0.0.123"
transformers = "^4.27.3"
rich = "^13.3.4"
functions-framework = "^3.3.0"
google-cloud-storage = "^2.8.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
mypy = "^1.1.1"
pyproject-flake8 = "^6.0.0.post1"
pytest = "^7.2.2"
ipykernel = "^6.22.0"
python-dotenv = "^1.0.0"
jupyterlab = "^3.6.2"

[tool.isort]
profile = "black"
line_length = 79

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
